calib/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
certifi==2020.6.20
cffi==1.15.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
freetype-py==2.2.0
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
notebook==6.4.8
numpy==1.19.5
open3d==0.9.0.0
opencv-contrib-python==4.5.5.64
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pycparser==2.21
pyglet==1.5.23
Pygments==2.11.2
PyMCubes==0.1.2
PyOpenGL==3.1.0
pyparsing==3.0.7
pyrender==0.1.45
pyrsistent==0.18.0
python-dateutil==2.8.2
pyzmq==22.3.0
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.6.0
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
trimesh==3.10.2
typing_extensions==4.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
